pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A Django GraphQL (Graphene) base template"
authors = ["dhilipsiva <dhilipsiva@pm.me>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
django-extensions = "^3.1.5"
graphene-django = "^2.15.0"
django-cors-headers = "^3.11.0"
psycopg2-binary = "^2.9.3"
boto3 = "^1.20.51"
celery = "^5.2.3"
django-graphql-jwt = "^0.3.4"

[tool.poetry.dev-dependencies]
neovim = "^0.3.1"
ipython = "^8.0.1"
flake8 = "^4.0.1"
pep8 = "^1.7.1"
ipdb = "^0.13.9"
black = "^22.1.0"
isort = "^5.10.1"
pip = "^22.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
# Make sure this matches `*.py` in .editorconfig
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
use_parentheses = true
